--- supriya/ugens/PseudoUGen.py
@@ -1,14 +1,0 @@
-import abc
-
-from supriya.system import SupriyaObject
-
-
-class PseudoUGen(SupriyaObject):
-
-    ### CLASS VARIABLES ###
-
-    ### INITIALIZER ###
-
-    @abc.abstractmethod
-    def __init__(self):
-        raise NotImplementedError
