+++ supriya/ugens/granular.py
@@ -1,7 +1,7 @@
 import collections
 
 from supriya import CalculationRate
-from supriya.synthdefs import MultiOutUGen
+from supriya.synthdefs import MultiOutUGen, UGen
 
 
 class GrainBuf(MultiOutUGen):
@@ -74,6 +74,32 @@ class GrainIn(MultiOutUGen):
             ("position", 0),
             ("envelope_buffer_id", -1),
             ("maximum_overlap", 512),
+        ]
+    )
+    _valid_calculation_rates = (CalculationRate.AUDIO,)
+
+
+class PitchShift(UGen):
+    """
+    A pitch shift unit generator.
+
+    ::
+
+        >>> source = supriya.ugens.SoundIn.ar()
+        >>> supriya.ugens.PitchShift.ar(
+        ...     source=source,
+        ...     )
+        PitchShift.ar()
+
+    """
+
+    _ordered_input_names = collections.OrderedDict(
+        [
+            ("source", None),
+            ("window_size", 0.2),
+            ("pitch_ratio", 1.0),
+            ("pitch_dispersion", 0.0),
+            ("time_dispersion", 0.0),
         ]
     )
     _valid_calculation_rates = (CalculationRate.AUDIO,)
