+++ supriya/ugens/osc.py
@@ -88,6 +88,23 @@ class Index(PureUGen):
     _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
 
 
+class LFCub(PureUGen):
+    """
+    A sine-like oscillator unit generator.
+
+    ::
+
+        >>> supriya.ugens.LFCub.ar()
+        LFCub.ar()
+
+    """
+
+    _ordered_input_names = collections.OrderedDict(
+        [("frequency", 440.0), ("initial_phase", 0.0)]
+    )
+    _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
+
+
 class LFGauss(PureUGen):
     """
     A non-band-limited gaussian function oscillator.
